Updated return type for world population so that it can return the proper number (if the query works) Updated test for world population to compare to the actual number

+++ src/main/java/com/napier/sem/App.java
@@ -165,7 +165,7 @@ public class App {
         return cityQueries.getCitiesInDistrictByLargestToSmallestPopulation(district);
     }
 
-    public int getWorldPopulation() {
+    public long getWorldPopulation() {
         WorldQueries worldQueries = new WorldQueries(dbDriver.getConn());
         return worldQueries.getWorldPopulation();
     }
